requirements.txt
cryptography~=41.0.3
pip~=23.2.1
amqp~=5.1.1
vine~=5.0.0
pytz~=2023.3.post1
Django~=4.2.5
kombu~=5.3.2
billiard~=4.1.0
redis~=5.0.0
requests~=2.31.0
celery~=5.3.4
click~=8.1.7
python-dateutil~=2.8.2
asgiref~=3.7.2
sqlparse~=0.4.4
colorama~=0.4.6
numpy~=1.24.4
django-allauth~=0.56.1
PyJWT~=2.8.0
certifi~=2023.7.22
tzlocal~=5.0.1
wcwidth~=0.2.6
cffi~=1.15.1
six~=1.16.0
oauthlib~=3.2.2
idna~=3.4
urllib3~=2.0.4
pycparser~=2.21
setuptools~=49.2.1
APScheduler~=3.10.4
django-filter~=23.3
djangorestframework~=3.14.0
django-modeltranslation~=0.18.11
news~=1.0